requirements.txt
Django==1.6.2
MySQL-python==1.2.5
argparse==1.2.1
cloudinary==1.0.14
distribute==0.6.34
dj-database-url==0.2.2
dj-static==0.0.5
django-crispy-forms==1.4.0
django-haystack==2.1.0
django-toolbelt==0.0.1
gunicorn==18.0
newrelic==2.12.0.10
psycopg2==2.5.2
pystache==0.5.3
static==1.0.2
wsgiref==0.1.2
